Optimized height property maintenance during Insert and Remove
Previously, the height function was called on every node along the path of an insertion or removal, as well as all nodes involded in a rotation. This was inneficient.

With this change, stored height data should be kept up-to-date by allowing each node to request the stored height data of its children. So while each node along the modification path will still update its height value, the height value can be determined in constant time which is much better.